PrimeRoot - v.2.1.p Pseudo-stable variable base prototype.
Work in progress bugs:
Number 18071751121 gives partial results then returns a segmentation fault

Changes:
Addressed the bug that was returning X and Y orders of magnitude higher than the actual result.

+++ include/VB_Iterator.h
@@ -13,6 +13,7 @@
 #include <boost/container/list.hpp>
 
 
+
 class VB_Iterator {
 
 private:
@@ -60,7 +61,7 @@ private:
     void checkResult(int position, mpz_t result);
 
     ///Step functions
-    void setFactor(uint8_t segment, uint8_t position, mpz_t number);
+    void setFactor(uint8_t segment, int position, mpz_t number);
     void getSubResult(uint8_t xSegment, uint8_t ySegment, mpz_t result, mpz_t previousResult);
     void resetNode(int position);
